Cleanup after merging gui with logic.
