--- algorithms/gac/gac.py
@@ -1,131 +1,0 @@
-__author__ = 'bompi88'
-
-import copy
-
-
-class GAC:
-    def __init__(self):
-        self.variable_map = {}  # Key: variable, Value: all constraints which includes this variable
-        self.constraint_map = {}  # Key: expression/constraint, Value: all variables in this constraint
-        self.queue = []  # tuples (variable, expression/constraint)
-        self.functions = {} # Key: expression/constraint, Value: method to evaluate constraint
-        self.domains = {}  # Key: variable, Value: List of available values that represent the domain of the variable
-
-    def initialize(self, variables, domains, constraints):
-
-        self.domains = domains
-
-        for c in constraints:
-            elements = c.split()
-
-            vars = list(set(elements) & set(variables))
-
-            self.constraint_map[c] = vars
-
-            for v in vars:
-                if v in self.variable_map.keys():
-                    self.variable_map[v].append(c)
-                else:
-                    self.variable_map[v] = [c]
-
-                self.queue.append((v, c))
-
-            self.functions[c] = self.makefunc(vars, c)
-
-    def domain_filtering(self):
-        while len(self.queue):
-            todo_revise = self.queue.pop(0)
-
-            if self.revise_star(todo_revise):
-                self.add_to_queue(todo_revise, True)
-
-        return self.domains
-
-    def revise_star(self, cv_pair):
-        # todo: add multiple variable support
-
-        size_domain_old = len(self.domains[cv_pair[0]])
-
-        vars_to_consider = copy.deepcopy(self.constraint_map[cv_pair[1]])
-        vars_to_consider.remove(cv_pair[0])
-
-        to_remove = []
-
-        # For all values in focus variable's domain
-        for d1 in self.domains[cv_pair[0]]:
-            retain = False
-
-            if vars_to_consider and len(vars_to_consider) > 0:
-                for d2 in self.domains[vars_to_consider[0]]:
-                    var_order = cv_pair[1].split()
-
-                    if var_order[0] is cv_pair[0]:
-                        if self.functions[cv_pair[1]](d1, d2):
-                            retain = True
-                    else:
-                        if self.functions[cv_pair[1]](d2, d1):
-                            retain = True
-            else:
-                if self.functions[cv_pair[1]](d1):
-                    retain = True
-
-            if not retain and d1 in self.domains[cv_pair[0]]:
-                to_remove.append(d1)
-
-        # Remove the values from the domain if
-        for r in to_remove:
-            self.domains[cv_pair[0]].remove(r)
-
-        if len(self.domains[cv_pair[0]]) is 1:
-            for el in self.queue:
-                if el[0] is cv_pair[0]:
-                    self.queue.remove(el)
-
-        size_domain_new = len(self.domains[cv_pair[0]])
-
-        # if domain gets reduced, return true
-        if size_domain_new < size_domain_old:
-            return True
-        else:
-            return False
-
-    def rerun(self, assumption):
-        self.domains[assumption[0]] = assumption[1]
-        self.add_to_queue((assumption, None))
-        self.domain_filtering()
-
-    def add_to_queue(self, objective, exclude=False):
-        # all affected variables
-        variables = set()
-
-        # All constraints that has to be rechecked
-        constraints = set()
-
-        # All constraints given the current variable
-
-        for c in self.variable_map[objective[0]]:
-            constraints.add(c)
-
-        # Get all affected variables
-        for c in constraints:
-            for v in self.constraint_map[c]:
-                variables.add(v)
-
-        for v in variables:
-            cs = self.variable_map[v]
-
-            for c in cs:
-                # Do not consider the last revised constraint
-                if exclude and c is objective[1]:
-                    continue
-                else:
-                    self.queue.append((v, c))
-
-    @staticmethod
-    def makefunc(var_names, expression, envir=globals()):
-        args = ""
-
-        for n in var_names:
-            args = args + "," + n
-
-        return eval("(lambda " + args[1:] + ": " + expression + ")", envir)
